Удаление и редактирование Проектанта только Методисту. Дизайн Groups, Вывод групп с учетом входа

+++ IS-technopark/Account/Info.aspx.cs
@@ -11,6 +11,7 @@ using System.Drawing;
 using System.Text;
 using System.Threading.Tasks;
 using System.Data.SqlClient;
+using IS_technopark.Account;
 
 namespace IS_technopark
 {
@@ -23,7 +24,6 @@ namespace IS_technopark
         DataTable table = new DataTable();
         string id_l = "";
         List<string> id_l_list = new List<string>();
-        //string sqlconnection = "Data Source=(DESCRIPTION =(ADDRESS_LIST=(ADDRESS=(PROTOCOL=TCP)(HOST = 127.0.0.1)(PORT = 1521)))(CONNECT_DATA =(SERVICE_NAME = xe)));User ID=TECHNOPARK;Password=DIP1937;";
 
         protected void Page_Load(object sender, EventArgs e)
         {
@@ -44,7 +44,6 @@ namespace IS_technopark
                         oraReader.GetValues(values);
                         id_l_list.Add(values[0].ToString());
                     }
-
                 }
                 a += 1;
                 // Response.Write(cb + "<b>tut</b><br/>");
@@ -58,52 +57,58 @@ namespace IS_technopark
             GridView1.EditIndex = -1;
             GridView1.DataBind();
             if (this.IsPostBack)
-            {
-                string command = Technopark.SelectCommand;
-                Technopark.SelectCommand = "SELECT * FROM TECHNOPARK.LEARNER WHERE (FIO LIKE '%" + TextBox1.Text + "%') and ID_LEARNER!=0";
-                Technopark.DataBind();
-                GridView1.DataBind();
-            }
-
+                {
+                    string command = Technopark.SelectCommand;
+                    Technopark.SelectCommand = "SELECT * FROM TECHNOPARK.LEARNER WHERE (FIO LIKE '%" + TextBox1.Text + "%') and ID_LEARNER!=0";
+                    Technopark.DataBind();
+                    GridView1.DataBind();
+                }
         }
 
         protected void GridView1_RowUpdating(object sender, GridViewUpdateEventArgs e)
         {
-            if (this.IsPostBack)
+            if (Class_FIO.Employees_position == "1")
             {
-                string command = Technopark.SelectCommand;
-                Technopark.SelectCommand = "SELECT * FROM TECHNOPARK.LEARNER WHERE (FIO LIKE '%" + TextBox1.Text + "%') and ID_LEARNER!=0";
-                Technopark.DataBind();
-                GridView1.DataBind();
-            }
-            try
-            {
-                using (OracleConnection oraclelcon = new OracleConnection("Data Source =127.0.0.1:1521/xe; User ID =Technopark;  password = DIP1937;"))
+                if (this.IsPostBack)
                 {
-                    GridViewRow row = GridView1.Rows[e.RowIndex];
-                    string query = "UPDATE TECHNOPARK.LEARNER SET FIO=:FIO, CLASS=:CLASS, BIRTHDAY=:BIRTHDAY, SCHOOL=:SCHOOL, PHONE=:PHONE, SHIFT=:SHIFT, E_MAIL=:E_MAIL, INTERESTS=:INTERESTS, COMMENTS=:COMMENTS WHERE ID_LEARNER=:ID_LEARNER";
-                    OracleCommand oraclecmd = new OracleCommand(query, oraConnection);
-                    oraclelcon.Open();
-                    oraclecmd.Connection.Open();
-                    oraclecmd.Parameters.Add("FIO", e.NewValues["FIO"]);
-                    oraclecmd.Parameters.Add("CLASS", e.NewValues["CLASS"]);
-                    oraclecmd.Parameters.Add("BIRTHDAY", Convert.ToDateTime(e.NewValues["BIRTHDAY"]));
-                    oraclecmd.Parameters.Add("SCHOOL", e.NewValues["SCHOOL"]);
-                    oraclecmd.Parameters.Add("PHONE", e.NewValues["PHONE"]);
-                    oraclecmd.Parameters.Add("SHIFT", e.NewValues["SHIFT"]);
-                    oraclecmd.Parameters.Add("E_MAIL", e.NewValues["E_MAIL"]);
-                    oraclecmd.Parameters.Add("INTERESTS", e.NewValues["INTERESTS"]);
-                    oraclecmd.Parameters.Add("COMMENTS", e.NewValues["COMMENTS"]);
-                    oraclecmd.Parameters.Add("ID_LEARNER", GridView1.DataKeys[e.RowIndex].Value);
+                    string command = Technopark.SelectCommand;
+                    Technopark.SelectCommand = "SELECT * FROM TECHNOPARK.LEARNER WHERE (FIO LIKE '%" + TextBox1.Text + "%') and ID_LEARNER!=0";
+                    Technopark.DataBind();
+                    GridView1.DataBind();
+                }
+                try
+                {
+                    using (OracleConnection oraclelcon = new OracleConnection("Data Source =127.0.0.1:1521/xe; User ID =Technopark;  password = DIP1937;"))
+                    {
+                        GridViewRow row = GridView1.Rows[e.RowIndex];
+                        string query = "UPDATE TECHNOPARK.LEARNER SET FIO=:FIO, CLASS=:CLASS, BIRTHDAY=:BIRTHDAY, SCHOOL=:SCHOOL, PHONE=:PHONE, SHIFT=:SHIFT, E_MAIL=:E_MAIL, INTERESTS=:INTERESTS, COMMENTS=:COMMENTS WHERE ID_LEARNER=:ID_LEARNER";
+                        OracleCommand oraclecmd = new OracleCommand(query, oraConnection);
+                        oraclelcon.Open();
+                        oraclecmd.Connection.Open();
+                        oraclecmd.Parameters.Add("FIO", e.NewValues["FIO"]);
+                        oraclecmd.Parameters.Add("CLASS", e.NewValues["CLASS"]);
+                        oraclecmd.Parameters.Add("BIRTHDAY", Convert.ToDateTime(e.NewValues["BIRTHDAY"]));
+                        oraclecmd.Parameters.Add("SCHOOL", e.NewValues["SCHOOL"]);
+                        oraclecmd.Parameters.Add("PHONE", e.NewValues["PHONE"]);
+                        oraclecmd.Parameters.Add("SHIFT", e.NewValues["SHIFT"]);
+                        oraclecmd.Parameters.Add("E_MAIL", e.NewValues["E_MAIL"]);
+                        oraclecmd.Parameters.Add("INTERESTS", e.NewValues["INTERESTS"]);
+                        oraclecmd.Parameters.Add("COMMENTS", e.NewValues["COMMENTS"]);
+                        oraclecmd.Parameters.Add("ID_LEARNER", GridView1.DataKeys[e.RowIndex].Value);
 
-                    oraclecmd.ExecuteNonQuery();
-                    GridView1.EditIndex = -1;
+                        oraclecmd.ExecuteNonQuery();
+                        GridView1.EditIndex = -1;
+                    }
+                }
+                catch
+                {
+                    Label1.Visible = true;
+                    Label1.Text = "Проверьте введенные данные!";
                 }
             }
-            catch
+            else
             {
-                Label1.Visible = true;
-                Label1.Text = "Проверьте введенные данные!";
+                Response.Write("<script>alert('Только методист может редактировать данные!')</script>");
             }
 
         }
@@ -176,37 +181,51 @@ namespace IS_technopark
 
         protected void GridView1_RowDeleting(object sender, GridViewDeleteEventArgs e)
         {
-            try
+            if (Class_FIO.Employees_position == "1")
             {
-                using (OracleConnection oraclelcon = new OracleConnection("Data Source =127.0.0.1:1521/xe; User ID =Technopark;  password = DIP1937;"))
+                try
                 {
+                  using (OracleConnection oraclelcon = new OracleConnection("Data Source =127.0.0.1:1521/xe; User ID =Technopark;  password = DIP1937;"))
+                  {
+                        oraclelcon.Open();
+                        GridViewRow row = GridView1.Rows[e.RowIndex];
+                        string query = "DELETE FROM TECHNOPARK.LEARNER WHERE ID_LEARNER = :ID_LEARNER";
+                        OracleCommand oraclecmd = new OracleCommand(query, oraConnection);
+                        oraclecmd.Connection.Open();
+                        oraclecmd.Parameters.Add("ID_LEARNER", GridView1.DataKeys[e.RowIndex].Value);
 
-                    //GridViewRow row = GridView1.Rows[e.RowIndex];
-                    string query = "DELETE FROM TECHNOPARK.LEARNER WHERE ID_LEARNER = :ID_LEARNER";
-                    OracleCommand oraclecmd = new OracleCommand(query, oraConnection);
-                    oraclelcon.Open();
-                    oraclecmd.Connection.Open();
-                    oraclecmd.ExecuteNonQuery();
-
+                        oraclecmd.ExecuteNonQuery();
+                    }
+                }
+                catch
+                {
+                    Response.Write("<script>alert('Eeegc')</script>");
                 }
             }
-            catch
+            else
             {
-
+                Response.Write("<script>alert('Только методист может удалять данные!')</script>");
             }
         }
 
         protected void GridView1_RowEditing(object sender, GridViewEditEventArgs e)
         {
-            if (this.IsPostBack)
+            if (Class_FIO.Employees_position == "1")
             {
-                string command = Technopark.SelectCommand;
-                Technopark.SelectCommand = "SELECT * FROM TECHNOPARK.LEARNER WHERE (FIO LIKE '%" + TextBox1.Text + "%') and ID_LEARNER!=0";
-                Technopark.DataBind();
+                    if (this.IsPostBack)
+                {
+                    string command = Technopark.SelectCommand;
+                    Technopark.SelectCommand = "SELECT * FROM TECHNOPARK.LEARNER WHERE (FIO LIKE '%" + TextBox1.Text + "%') and ID_LEARNER!=0";
+                    Technopark.DataBind();
+                    GridView1.DataBind();
+                }
+                GridView1.EditIndex = e.NewEditIndex;
                 GridView1.DataBind();
             }
-            GridView1.EditIndex = e.NewEditIndex;
-            GridView1.DataBind();
+            else
+            {
+                Response.Write("<script>alter('EDIT');</script>");
+            }
         }
 
         protected void Button2_Click(object sender, EventArgs e)
@@ -218,7 +237,7 @@ namespace IS_technopark
         {
             oraConnection.Open();
             string command = Technopark.SelectCommand;
-            Technopark.SelectCommand = "SELECT * FROM TECHNOPARK.LEARNER";
+            Technopark.SelectCommand = "SELECT * FROM TECHNOPARK.LEARNER order by FIO";
             Technopark.DataBind();
             GridView1.DataBind();
             oraConnection.Close();
